- widened ttl variable in particle_t to short to allow longer living particles than 6 seconds. (patch by Graf Zahl)

+++ src/p_effect.h
@@ -57,7 +57,7 @@ struct particle_t
 	fixed_t	x,y,z;
 	fixed_t velx,vely,velz;
 	fixed_t accx,accy,accz;
-	BYTE	ttl;
+	short	ttl;
 	BYTE	trans;
 	WORD	size;
 	BYTE	bright:1;
